Get the dcat data service.

Get the dcat data service.

Usage

i14y_get_data_service(id = NULL, language = "de")

Arguments

id

string. The id of the dcat data service.

language

string. The language of the response data.

Value

a list

Examples

i14y_get_data_service(
 id = "0aaed69d-15ff-415a-8a38-4f9f0c3b429e",
 language = "en"
)
#> $accessRights
#> $accessRights$cultureCode
#> [1] "en"
#> 
#> $accessRights$text
#> [1] "Public"
#> 
#> 
#> $conformsTo
#>                                  href label.cultureCode       label.text
#> 1 https://developers.arcgis.com/rest/                en ArcGIS REST APIs
#> 
#> $contactPoint
#>          child       emailInternet note org     telWorkVoice
#> 1 Organization webgis@swisstopo.ch   NA  NA +41 58 469 01 11
#>   adrWork.cultureCode                               adrWork.text fn.cultureCode
#> 1                  en Seftigenstrasse 264\nP.O. Box\n3084 Wabern             en
#>                                  fn.text
#> 1 Federal Office of Topography swisstopo
#> 
#> $datasetLinks
#> list()
#> 
#> $description
#> $description$cultureCode
#> [1] "en"
#> 
#> $description$text
#> [1] "The Identify service can be used to discover features at a specific location. Queryable data can additionally be filtered by attributes. The Identify service can be used for Reverse Geocoding (the process of converting a location as described by geographic coordinates to an address or a place name) operations."
#> 
#> 
#> $documentation
#>                                                                               href
#> 1           https://api.geo.admin.ch/api/faq/index.html#which-layers-are-queryable
#> 2             https://api.geo.admin.ch/services/sdiservices.html#identify-features
#> 3                           http://api.geo.admin.ch/services/sdiservices.html#id11
#> 4 https://api.geo.admin.ch/services/sdiservices.html#examples-of-reverse-geocoding
#> 5          https://api.geo.admin.ch/api/faq/index.html#which-layers-have-a-tooltip
#>   label.cultureCode        label.text
#> 1                en    Queryable data
#> 2                en           API Doc
#> 3                en          Examples
#> 4                en Reverse Geocoding
#> 5                en    Available data
#> 
#> $endpointDescription
#>                                                                   href
#> 1 https://api.geo.admin.ch/services/sdiservices.html#identify-features
#>   label.cultureCode label.text
#> 1                en    API Doc
#> 
#> $endpointUrl
#>                                                             href label
#> 1 https://api.geo.admin.ch/rest/services/api/MapServer/identify?    NA
#> 
#> $id
#> [1] "0aaed69d-15ff-415a-8a38-4f9f0c3b429e"
#> 
#> $keyword
#>   cultureCode
#> 1          en
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  text
#> 1 Addresses, Administrative units, Agricultural and aquaculture facilities, Area management/restriction/regulation zones and reporting units, Atmospheric conditions, Bio-geographical regions, Buildings, Cadastral parcels, Coordinate reference systems, Elevation, Energy resources, Environmental monitoring facilities, Geographical grid systems, Geographical names, Geology, Habitats and biotopes, Human health and safety, Hydrography, Land cover, Land use, Meteorological geographical features, Mineral resources, Natural risk zones, Orthoimagery, Population distribution ? demography, Production and industrial facilities, Protected sites, Soil, Species distribution, Statistical units, Transport networks, Utility and governmental services
#> 
#> $landingPage
#>                                                                   href
#> 1 https://api.geo.admin.ch/services/sdiservices.html#identify-features
#>   label.cultureCode label.text
#> 1                en    API Doc
#> 
#> $license
#> NULL
#> 
#> $nextVersions
#> list()
#> 
#> $previousVersion
#> NULL
#> 
#> $publisher
#> $publisher$name
#> $publisher$name$cultureCode
#> [1] "en"
#> 
#> $publisher$name$text
#> [1] "Federal Office of Topography swisstopo"
#> 
#> 
#> $publisher$type
#> NULL
#> 
#> 
#> $theme
#>   cultureCode           text
#> 1          en Geoinformation
#> 
#> $title
#> $title$cultureCode
#> [1] "en"
#> 
#> $title$text
#> [1] "Identify Features Service - Federal Spatial Data Infrastructure"
#> 
#> 
#> $version
#> NULL
#> 
#> $versionNotes
#> NULL
#>