Get dataset distributions by identifier

Get dataset distributions by identifier

Usage

i14y_get_dataset_distributions(identifier = NULL, language = "de")

Arguments

identifier

string. The identifier of the dcat dataset.

language

string. The language of the response data.

Value

a list

Examples

i14y_get_dataset_distributions(identifier = "HCL_NOGA",  language = "en")
#>                                 accessUrl availability byteSize checksum
#> 1 https://www.kubb-tool.bfs.admin.ch/, NA           NA       NA       NA
#>   conformsTo coverage documentation downloadUrl
#> 1       NULL     NULL          NULL        NULL
#>                                     id identifier image       language license
#> 1 d648590f-37cc-45aa-a1bf-99007efc45cc         NA  NULL de, fr, it, en      NA
#>   mediaType modified packagingFormat               releaseDate rights
#> 1        NA       NA              NA 2008-01-01T00:00:00+01:00     NA
#>   spatialResolution status temporalResolution description.cultureCode
#> 1              NULL   NULL                 NA                      en
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            description.text
#> 1 The General Classification of Economic Activities (NOGA) presents the universe of economic activities partitioned in such a way that a NOGA code can be associated with a statistical unit carrying them out. The NOGA is the national version of the Statistical classification of economic activities in the European Community (NACE). The NOGA is structurally linked to NACE for the third and fourth levels. NACE is structurally linked to the International Standard Industrial Classification of All Economic Activities (ISIC) for the first and second levels.
#>   format.cultureCode format.text title.cultureCode
#> 1                 en        HTML                en
#>                                             title.text
#> 1 General Classification of Economic Activities – NOGA