Get Concept view entity by ID

Get Concept view entity by ID

Usage

i14y_get_concept(id = NULL, language = "de")

Arguments

id

string. The Id of the response data.

language

string. The language of the response data.

Value

a list

Examples

i14y_get_concept(
 id = "08d94604-e058-62a2-aa25-53f84b974201", # DV_NOGA_DIVISION
 language = "en"
)
#> $agencyName
#> $agencyName$cultureCode
#> [1] "en"
#> 
#> $agencyName$text
#> [1] "Federal Statistical Office"
#> 
#> 
#> $codelistEntryValueMaxLength
#> [1] 2
#> 
#> $codeListEntryValueType
#> [1] "String"
#> 
#> $codeListId
#> [1] "08d9460c-a59e-6824-a442-96d195a63d58"
#> 
#> $conceptType
#> [1] "CodeList"
#> 
#> $conformsTo
#> list()
#> 
#> $description
#> $description$cultureCode
#> [1] "en"
#> 
#> $description$text
#> [1] "Second level of the General Classification of Economic Activities (NOGA) composed of headings identified by a two-digit numerical code. This second level is determined by the International Standard Industrial Classification of All Economic Activities (ISIC Rev.4)."
#> 
#> 
#> $id
#> [1] "08d94604-e058-62a2-aa25-53f84b974201"
#> 
#> $identifier
#> [1] "DV_NOGA_DIVISION"
#> 
#> $keywords
#> list()
#> 
#> $maxLength
#> NULL
#> 
#> $maxValue
#> NULL
#> 
#> $measurementUnit
#> NULL
#> 
#> $minLength
#> NULL
#> 
#> $minValue
#> NULL
#> 
#> $name
#> $name$cultureCode
#> [1] "en"
#> 
#> $name$text
#> [1] "NOGA Division"
#> 
#> 
#> $nbDecimal
#> NULL
#> 
#> $pattern
#> NULL
#> 
#> $responsibleDeputy
#> [1] "Costantini Sergio BFS"
#> 
#> $responsiblePerson
#> [1] "Froidevaux Marc BFS"
#> 
#> $themes
#> list()
#> 
#> $validFrom
#> [1] "2007-12-31T23:00:00+00:00"
#> 
#> $validTo
#> NULL
#> 
#> $version
#> [1] "3.0.0"
#>