Get dataset description by identifier

Get dataset description by identifier

i14y_get_dataset_description(identifier = NULL, language = "de")

Arguments

identifier

string. The identifier of the dcat dataset.

language

string. The language of the response data.

Value

a list

Examples

i14y_get_dataset_description(identifier = "HCL_NOGA",  language = "en")
#> $accessRights
#> $accessRights$cultureCode
#> [1] "en"
#> 
#> $accessRights$text
#> [1] "Public"
#> 
#> 
#> $confidentialityPerson
#> NULL
#> 
#> $conformsTo
#>                                     href label.cultureCode
#> 1 https://www.ech.ch/de/ech/ech-0115/1.0                en
#>                                                                           label.text
#> 1 eCH-0115 Data standard NOGA General Classification of Economic Activities (DE, FR)
#> 
#> $contactPoint
#>          child     emailInternet org     telWorkVoice adrWork.cultureCode
#> 1 Organization noga@bfs.admin.ch  NA +41 58 463 65 23                  en
#>                                                                     adrWork.text
#> 1 Business Registers Data URD\nEspace de l'Europe 10\nCH-2010 Neuchâtel\nSchweiz
#>   fn.cultureCode                    fn.text note.cultureCode
#> 1             en Federal Statistical Office               en
#>                                      note.text
#> 1 Monday to Friday\n8.30–11.30 and 14.00–16.00
#> 
#> $description
#> $description$cultureCode
#> [1] "en"
#> 
#> $description$text
#> [1] "The General Classification of Economic Activities (NOGA) is the Swiss version of the European classification of economic activities NACE. It has been defined at European level in the different official languages of the Member States taking into account the economic structure of the majority of countries up to the fourth position. The NOGA was developed by the FSO in collaboration with experts from the public administration and numerous umbrella organisations. The particularities and/or important activities of the national economy have been incorporated in the last two positions of the six-digit code."
#> 
#> 
#> $documentation
#> list()
#> 
#> $frequency
#> NULL
#> 
#> $geoIvId
#> list()
#> 
#> $id
#> [1] "4f71b53b-8e67-46ca-90bf-c5fef19a9db0"
#> 
#> $identifier
#> [1] "HCL_NOGA"
#> 
#> $image
#> list()
#> 
#> $isReferencedBy
#> list()
#> 
#> $issued
#> [1] "2008-01-01T00:00:00+01:00"
#> 
#> $keyword
#>   cultureCode                  text
#> 1          en                Prices
#> 2          en       Work and income
#> 3          en                  NOGA
#> 4          en      National economy
#> 5          en Industry and services
#> 6          en             Registers
#> 7          en   Economic Activities
#> 
#> $landingPage
#>                        href label
#> 1 http://noga.bfs.admin.ch/    NA
#> 
#> $language
#> [1] "de" "fr" "it" "en"
#> 
#> $modified
#> NULL
#> 
#> $nextVersions
#> list()
#> 
#> $otherIdentifier
#> list()
#> 
#> $previousVersion
#> NULL
#> 
#> $publisher
#> $publisher$name
#> $publisher$name$cultureCode
#> [1] "en"
#> 
#> $publisher$name$text
#> [1] "Federal Statistical Office"
#> 
#> 
#> $publisher$type
#> NULL
#> 
#> 
#> $qualifiedAttribution
#> list()
#> 
#> $qualifiedAttributionComplement
#> NULL
#> 
#> $qualifiedRelation
#> list()
#> 
#> $relation
#> list()
#> 
#> $retentionPeriod
#> NULL
#> 
#> $retentionPeriodDescription
#> NULL
#> 
#> $spatial
#> list()
#> 
#> $temporalCoverage
#> list()
#> 
#> $theme
#>   cultureCode                text
#> 1          en             Economy
#> 2          en Official statistics
#> 3          en          Businesses
#> 
#> $title
#> $title$cultureCode
#> [1] "en"
#> 
#> $title$text
#> [1] "General Classification of Economic Activities – NOGA"
#> 
#> 
#> $type
#> NULL
#> 
#> $version
#> NULL
#> 
#> $versionNotes
#> NULL
#>